requirements.txt
torch
diffusers
transformers
accelerate
pillow
